refactor: name of pages
